国科大程序适配

+++ fan_control_code_adapt_c/include/message.h
@@ -170,7 +170,7 @@ message_data CmdPlaneTrajectory {
 message_data CmdPlanePower {
     uint8_t device_id; // 3    设备号 内容：01/02/03……
     uint8_t cmd_type; // 4    指令类型 内容：0x12
-    uint8_t cmd_data; // 5    指令内容 内容：0x00 执行中 0x01 停机
+    uint32_t cmd_data; // 5    指令内容 内容：0x00 执行中 0x01 停机
 };
 
 message_data FanTestData{
